Improve build and load scripts.
build.sh:
  Two command line switches:
  -c Clean thoroughly before building
  -l Call pscnv_load.sh after building

pscnv_load.sh
  Environment variables to determines GPU application to run for reclocking.
  Environment variable to override the default nvafakebios parameters
  (i.e. offsets of power state to invalidate).
  Cosmetic fixes.

+++ build.sh
@@ -3,8 +3,8 @@
 # -u: Treat unset variables as an error when substituting.
 
 # Builds pscnv and libpscnv. Currently doesn't install anything anywhere,
-# but calls switch-to-pscnv.sh to load up the produced module pscnv.ko
-# directly from the build folder
+# but optionally (-s command line switch) calls pscnv_load.sh to load up
+# the produced module pscnv.ko directly from the build folder
 
 # INPUT VARIABLES
 # ===============
@@ -13,12 +13,29 @@
 # in build directory:
 # $PSCNV_BUILD_DIR   (default $PSCNV_SOURCE_DIR/build )
 
+# COMMAND LINE
+# ============
+# -c Clean thoroughly before building
+# -l Call pscnv_load.sh after building
+
+DOCLEAN=0
+DOLOAD=0
+while getopts "cl" opt; do
+    case "$opt" in
+    c)   DOCLEAN=1
+         ;;
+    l)   DOLOAD=1
+         ;;
+    esac
+done
+
 # SCRIPT BODY
 # ==========
 
 # Search "man bash" for "Assign Default Values" to understand these lines:
 : ${PSCNV_SOURCE_DIR:="$HOME/pscnv"}
 : ${PSCNV_BUILD_DIR:="$PSCNV_SOURCE_DIR/build"}
+
 echo -n "$(tput bold)"
 echo    "~~~ Dumping input variables and performing some checks..."
 echo -n "$(tput sgr0)"
@@ -32,29 +49,31 @@ SRC_DIR="$PSCNV_SOURCE_DIR"
 BUILD_DIR="$PSCNV_BUILD_DIR"
 
 # Clean old stuff up
-echo
-echo "$(tput bold)~~~ Cleaning $SRC_DIR/pscnv (make clean; make distclean)...$(tput sgr0)"
-cd "$SRC_DIR/pscnv"
-make clean
-make distclean
+if [[ $DOCLEAN == 1 ]]; then
+    echo
+    echo "$(tput bold)~~~ Cleaning $SRC_DIR/pscnv (make clean; make distclean)...$(tput sgr0)"
+    cd "$SRC_DIR/pscnv"
+    make clean
+    make distclean
 
-#echo
-#echo "$(tput bold)~~~ Cleaning $SRC_DIR/libpscnv (make clean)...$(tput sgr0)"
-#cd "$SRC_DIR/libpscnv"
-#make clean
+    echo
+    echo "$(tput bold)~~~ Cleaning $SRC_DIR/libpscnv (make clean)...$(tput sgr0)"
+    cd "$SRC_DIR/libpscnv"
+    make clean
 
-#echo
-#echo "$(tput bold)~~~ Cleaning $SRC_DIR/test (make clean)...$(tput sgr0)"
-#cd "$SRC_DIR/test"
-#make clean
+    echo
+    echo "$(tput bold)~~~ Cleaning $SRC_DIR/test (make clean)...$(tput sgr0)"
+    cd "$SRC_DIR/test"
+    make clean
 
-echo
-echo "$(tput bold)~~~ Removing build dir $BUILD_DIR...$(tput sgr0)"
-rm -rf "$BUILD_DIR"
+    echo
+    echo "$(tput bold)~~~ Removing $BUILD_DIR...$(tput sgr0)"
+    rm -rf "$BUILD_DIR"
+fi
 
 # Recreate build directory
 echo
-echo "$(tput bold)~~~ Creating and entering $BUILD_DIR...$(tput sgr0)"
+echo "$(tput bold)~~~ Entering $BUILD_DIR...$(tput sgr0)"
 mkdir -pv "$BUILD_DIR"
 cd "$BUILD_DIR"
 
@@ -80,19 +99,24 @@ echo
 echo -n "$(tput bold)$(tput setaf 2)"
 echo "~~~ Success! Pscnv was built!"
 echo -n "$(tput sgr0)"
-echo "To install and load into kernel: "
-echo "    sudo mkdir -p \"/lib/modules/\$(uname -r)/extra\""
-echo "    sudo cp \"$PSCNV_KO_PATH\" \"/lib/modules/\$(uname -r)/extra\""
-echo "    sudo depmod"
-echo "    sudo modprobe pscnv"
-echo
-echo "To load into kernel directly without installing:"
-echo "    sudo modprobe -v drm"
-echo "    sudo modprobe -v drm_kms_helper"
-echo "    sudo modprobe -v video"
-echo "    sudo modprobe -v i2c-algo-bit"
-echo "    sudo insmod \"$PSCNV_KO_PATH\""
-echo
-echo "To load using a \"smart\" script that removes current GPU driver and performs reclocking:"
-echo "    \"$SRC_DIR/pscnv_load.sh\""
+
+if [[ $DOLOAD == 1 ]]; then
+    $SRC_DIR/pscnv_load.sh
+else
+    echo "To install and load into kernel: "
+    echo "    sudo mkdir -p \"/lib/modules/\$(uname -r)/extra\""
+    echo "    sudo cp \"$PSCNV_KO_PATH\" \"/lib/modules/\$(uname -r)/extra\""
+    echo "    sudo depmod"
+    echo "    sudo modprobe pscnv"
+    echo
+    echo "To load into kernel directly without installing:"
+    echo "    sudo modprobe -v drm"
+    echo "    sudo modprobe -v drm_kms_helper"
+    echo "    sudo modprobe -v video"
+    echo "    sudo modprobe -v i2c-algo-bit"
+    echo "    sudo insmod \"$PSCNV_KO_PATH\""
+    echo
+    echo "To (re)load using a smart script that can also do reclocking (-r):"
+    echo "    \"$SRC_DIR/pscnv_load.sh\"" # -r
+fi
